resolve conflict with cidse2025
